Remove utils.
